[update]: v1 updated

+++ lib/core/services/preferences/preferences_services.dart
@@ -1,29 +1,39 @@
+import 'package:events_ticket/data/models/preference_model.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 class PreferencesServices {
   final supabase = Supabase.instance.client;
 
-  // Fonction pour mettre à jour les préférences utilisateur
+  // Update user preferences
   Future<void> updateUserPreferences(
       String userId, List<String> preferenceIds) async {
+    if (userId.isEmpty || preferenceIds.isEmpty) {
+      print("User ID or preference IDs are empty.");
+      return;
+    }
+
     try {
-      // Supprime toutes les préférences existantes pour l'utilisateur
       await supabase.from('user_preferences').delete().eq('user_id', userId);
 
-      // Crée une liste d'entrées pour les nouvelles préférences
       final newPreferences = preferenceIds.map((preferenceId) {
         return {'user_id': userId, 'preference_id': preferenceId};
       }).toList();
 
-      // Insère les nouvelles préférences
-      await supabase.from('user_preferences').insert(newPreferences);
+      if (newPreferences.isNotEmpty) {
+        await supabase.from('user_preferences').insert(newPreferences);
+      }
     } catch (error) {
-      print("Erreur lors de la mise à jour des préférences: $error");
+      print("Error updating preferences: $error");
     }
   }
 
-  // Récupère la préférence par titre
-  Future<dynamic> getPreferenceByTitle(String title) async {
+  // Get preference by title
+  Future<PreferencesModel?> getPreferenceByTitle(String title) async {
+    if (title.isEmpty) {
+      print("Title is empty.");
+      return null;
+    }
+
     try {
       final response = await supabase
           .from('preferences')
@@ -32,19 +42,24 @@ class PreferencesServices {
           .single();
 
       if (response.isNotEmpty) {
-        return response;
+        return PreferencesModel.fromMap(response);
       } else {
-        print("Erreur lors de la récupération de la préférence");
+        print("Preference with title '$title' not found.");
         return null;
       }
     } catch (error) {
-      print("Erreur lors de la récupération de la préférence");
+      print("Error retrieving preference: $error");
       return null;
     }
   }
 
-  // Met à jour ou insère une préférence par ID, titre et icône
+  // Add or update a preference
   Future<void> setPreference(String id, String title, String icon) async {
+    if (id.isEmpty || title.isEmpty) {
+      print("ID or title is empty.");
+      return;
+    }
+
     try {
       final preferenceData = {
         'id': id,
@@ -52,16 +67,53 @@ class PreferencesServices {
         'icon': icon,
       };
 
-      final response = await supabase
-          .from('preferences')
-          .upsert(preferenceData)
-          .eq('id', id);
+      await supabase.from('preferences').upsert(preferenceData);
+    } catch (error) {
+      print("Error inserting/updating preference: $error");
+    }
+  }
 
-      if (response.error != null) {
-        print("Erreur lors de l'insertion/mise à jour de la préférence");
+  // Fetch all preferences
+  Future<List<PreferencesModel>> getAllPreferences() async {
+    try {
+      final response =
+          await supabase.from('preferences').select('id, title, icon');
+      if (response.isNotEmpty) {
+        return (response as List<dynamic>)
+            .map((data) => PreferencesModel.fromMap(data))
+            .toList();
+      } else {
+        return [];
       }
     } catch (error) {
-      print("Erreur lors de l'insertion/mise à jour de la préférence");
+      print("Error fetching all preferences: $error");
+      return [];
+    }
+  }
+
+  // Fetch user preferences
+  Future<List<PreferencesModel>> getUserPreferences(String userId) async {
+    if (userId.isEmpty) {
+      print("User ID is empty.");
+      return [];
+    }
+
+    try {
+      final response = await supabase
+          .from('user_preferences')
+          .select('preference_id, preferences(id, title, icon)')
+          .eq('user_id', userId);
+
+      if (response.isNotEmpty) {
+        return (response as List<dynamic>)
+            .map((data) => PreferencesModel.fromMap(data['preferences']))
+            .toList();
+      } else {
+        return [];
+      }
+    } catch (error) {
+      print("Error fetching user preferences: $error");
+      return [];
     }
   }
 }
